requirements.txt
psycopg2==2.8.6
Django==3.1.5
djangorestframework==3.12.2
remote-pdb==2.1.0
faker==5.5.0
factory_boy==3.2.0